alx_travel_app/alx_travel_app/requirement.txt
amqp==5.3.1
asgiref==3.9.1
asttokens==3.0.0
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.1
drf-yasg==1.21.10
executing==2.2.1
inflection==0.5.1
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kombu==5.5.4
matplotlib-inline==0.1.7
mysql-connector-python==9.4.0
mysqlclient==2.2.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
rabbitmq==0.2.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
